second commit

+++ src/main/java/com/example/IgorWebApp30/controllers/AdminController.java
@@ -1,21 +1,33 @@
 package com.example.IgorWebApp30.controllers;
 
+import com.example.IgorWebApp30.model.Role;
 import com.example.IgorWebApp30.model.User;
+import com.example.IgorWebApp30.repository.UserRepository;
+import com.example.IgorWebApp30.service.RoleService;
+import com.example.IgorWebApp30.service.RoleServiceImpl;
 import com.example.IgorWebApp30.service.UserService;
 import com.example.IgorWebApp30.service.UserServiceImpl;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
+import java.util.Set;
 
 @Controller
 @RequestMapping("/admin")
 public class AdminController {
 
+    private final RoleService roleService;
+    private final UserServiceImpl userService;
+
     @Autowired
-    private UserServiceImpl userService;
+    public AdminController(UserServiceImpl userService, RoleService roleService) {
+        this.userService = userService;
+        this.roleService = roleService;
+    }
 
     @GetMapping
     public String adminPage(Principal principal, Model model) {
@@ -24,45 +36,51 @@ public class AdminController {
         return "admin";
     }
 
-    // Список всех пользователей
     @GetMapping("/users")
     public String listUsers(Model model) {
         model.addAttribute("users", userService.getAllUsers());
         return "list";
     }
 
-    // Форма для создания нового пользователя
     @GetMapping("/users/new")
     public String createUserForm(Model model) {
         model.addAttribute("user", new User());
+        model.addAttribute("roles", roleService.getAllRoles());
         return "create";
     }
 
-    // Сохранение нового пользователя
     @PostMapping("/users")
     public String saveUser(@ModelAttribute("user") User user) {
         userService.saveUser(user);
         return "redirect:/admin/users";
     }
 
-    // Форма для редактирования пользователя
     @GetMapping("/users/edit/{id}")
     public String editUserForm(@PathVariable Long id, Model model) {
-        model.addAttribute("user", userService.getUserById(id));
+        User user = userService.getUserById(id);
+        Set<Role> roles = roleService.getAllRoles();
+        model.addAttribute("user", user);
+        model.addAttribute("roles", roles);
         return "edit";
     }
 
-    // Обновление пользователя
     @PostMapping("/users/{id}")
     public String updateUser(@PathVariable Long id, @ModelAttribute("user") User user) {
         userService.updateUser(user);
         return "redirect:/admin/users";
     }
 
-    // Удаление пользователя
     @GetMapping("/users/delete/{id}")
     public String deleteUser(@PathVariable Long id) {
+        User user = userService.getUserById(id);
+        if(user == null) {
+            throw new RuntimeException("User not found");
+        }
+        user.getRoles().clear();
+        userService.saveUser(user);
         userService.deleteUser(id);
         return "redirect:/admin/users";
     }
+
+
 }
